requirements.txt
pandas
numpy
confluent-kafka==1.7.0
python-dotenv==0.17.1